feat: add additional process exec functions
This is a port of the utils/run.js file from cli-style:
https://github.com/dhis2/cli-style/blob/2bfb5ef5ee2b26ff7e27594fe36cef9eb12b2360/src/utils/run.js

It introduces cross-spawn as the default way to spawn processes which
works well on other platforms (e.g. windows) and exposes additional
functions to spawn and use the result of processes.

Notably, it adds the bin() function that resolves a command to the
nearest node_modules/.bin folder based on the cwd.

This is needed for cli-style, but also now needed for cli-utils-cypress,
so makes sense to generalize it here.