feat(reporter): add reporter.log function
Introduce a new logging function called reporter.log that is not printed
in standard operation, but visible with --verbose.

Helpful for when more information may be provided but is not of high
enough importance to be logged with the standard messages.

Changes the log levels to be integer based, to make it easier to print
all messages relevant up to a configured log level. E.g.:

-   with --debug, all messages from quiet, standard, verbose, and debug,
    are printed.

-   with --verbose, all messages from quiet, standard, and verbose, are
    printed.

-   with --quiet, only messages from the quiet log level are printed.