feat(loadmodule): add project-local module loader
The loadModule function will load modules from the project's
node_modules by default, and supports a preferLocal option that switches
the load function to use the parent's node_modules.

It is rare that a program using the loadModule helper to want to load a
module from the helper's node_modules, so passing in the parentModule
filename allows the helper to resolve using the parent's node_modules.

The options argument has the shape:

    {
        parentModule: <filepath>, // default: none
        preferLocal: <boolean>, // default: true
    }

Most often, the parentModule option can be __filename, but if you know
better feel free to set it to the absolute filepath that modules should
be resolved from.

The loadModule function returns a function when invoked:

    const loader = loadModule({
        parentModule: __filename,
    })

The loader function can then be used to load modules. It accepts a
single argument, and that is the moduleName:

    const mod = loader('module-a')

mod will then be the export of module-a from the node_modules that
belongs to the parentModule.

Bonus: This also works properly when using 'yarn link' to cli-helpers-engine.