feat(pretty-print): add a helper to print relative paths
This introduces a "pretty print" module that we can use to introduce
helpers for pretty print strings.

Printing relative file paths is something that comes up a lot. We want
to use the absolute path in the code as that makes everything more
robust when the cwd is manipulated or different from the expectation.

In those cases, it makes sense to have a function that just transforms
the absolute path to a relative path:

    pp.relativePath('/home/foo/bar', '/home')
    // output: foo/bar

The second argument defaults to process.cwd() and should be omitted for
normal use. We are interested in the user's current working directory
after all.

In cases where the user's cwd does not match what we want to show as
output, we can set the cwd to another path, such as the root project
directory to print src references relative to that.